Moved the creation of StreamApiLow to a method on PrivMXEndpoint instance
